requirements.txt
beautifulsoup4
certifi==2022.12.7
distlib==0.3.6
filelock==3.9.0
groq==0.11.0
packaging==23.2
Pillow==10.1.0
pipenv==2023.2.18
platformdirs==3.1.1
PyMuPDF
pygame==2.5.2
pyperclip==1.8.2
pytesseract==0.3.10
python-dotenv 
PyYAML==6.0.1
requests
virtualenv==20.20.0
virtualenv-clone==0.5.7


